[ refactor ] Move `LiftEx` to `level.rs`
