+++ src/syntax/core/redex.rs
@@ -1,7 +1,6 @@
-use crate::syntax::common::DBI;
-
-use super::level::LiftEx;
 use super::{CaseSplit, Closure, Neutral, Val, Variants};
+use crate::syntax::common::DBI;
+use crate::syntax::level::LiftEx;
 
 /// Reducible expressions.
 pub trait RedEx<T: Sized = Val>: Sized {
